chore(lint): adopt v2 lint-first selectors (N/B/SIM/UP/RUF/ERA/T20) + pyright strict on src/
- Add select = [E/F/W/I/N/B/SIM/UP/RUF/ERA/T20] to [tool.ruff.lint]
- Add [tool.pyright] block with basic mode + strict executionEnvironment for src/
- Add [tool.ruff.lint.per-file-ignores] covering camelCase coord API (N802/N803/N806/N811/N815/N816/N818), F401 side-effect imports, B017, T201, E402
- Fix all B904 (raise-in-except missing from), B905 (zip missing strict=), B007/RUF059 (unused unpacked vars, _ prefix), B018 (property access in raises)
- Fix SIM102/SIM103/SIM105/SIM108/SIM201, RUF005/RUF015/RUF034/RUF043/RUF046 findings
- Remove dead commented-out code (ERA001 true positives); noqa false-positive ERA001 in formula/section-header comments
- Suppress RUF012 for Enum member class (CharacterGroups) via per-file-ignores
- Add E501 to global ignore (line-length enforcement not desired project-wide)

+++ pd_book_tools/image_processing/cupy_processing/canvas.py
@@ -33,20 +33,20 @@ def map_content_onto_scaled_canvas_gpu(
     current_ratio = float(height) / float(width)
 
     if current_ratio >= height_width_ratio:
-        new_height = int(math.ceil(height / (1 - (whitespace_add * 2))))
-        new_width = int(math.ceil(new_height / height_width_ratio))
+        new_height = math.ceil(height / (1 - (whitespace_add * 2)))
+        new_width = math.ceil(new_height / height_width_ratio)
     else:
-        new_width = int(math.ceil(width / (1 - (whitespace_add * 2))))
-        new_height = int(math.ceil(new_width * height_width_ratio))
+        new_width = math.ceil(width / (1 - (whitespace_add * 2)))
+        new_height = math.ceil(new_width * height_width_ratio)
 
     canvas = cp.full((new_height, new_width), 255, dtype=cp.uint8)
 
     if force_align == Alignment.BOTTOM:
-        y_offset = new_height - (height + int(math.ceil(whitespace_add * new_height)))
+        y_offset = new_height - (height + math.ceil(whitespace_add * new_height))
     elif force_align == Alignment.CENTER:
         y_offset = int(new_height / 2) - int(height / 2)
     else:
-        y_offset = int(math.ceil(whitespace_add * new_height))
+        y_offset = math.ceil(whitespace_add * new_height)
 
     x_offset = int(new_width / 2) - int(width / 2)
